DMOH-10: Moved the not-direct to use Service Classes to their dedicated directories.
