+++ src/ChannelOpeningHoursHtmlServiceFactory.php
@@ -6,6 +6,7 @@ use StadGent\Services\OpeningHours\Client\ClientInterface;
 use Psr\SimpleCache\CacheInterface;
 use StadGent\Services\OpeningHours\Handler\Channel\OpeningHoursHtmlHandler;
 use StadGent\Services\OpeningHours\Handler\Channel\OpenNowHtmlHandler;
+use StadGent\Services\OpeningHours\Service\Channel\ChannelOpeningHoursHtmlService;
 
 /**
  * Factory to create the ChannelOpeningHoursHtmlService.
@@ -23,7 +24,7 @@ class ChannelOpeningHoursHtmlServiceFactory
      * @param \StadGent\Services\OpeningHours\Client\ClientInterface $client
      * @param \Psr\SimpleCache\CacheInterface $cache
      *
-     * @return \StadGent\Services\OpeningHours\ChannelOpeningHoursHtmlService
+     * @return \StadGent\Services\OpeningHours\Service\Channel\ChannelOpeningHoursHtmlService
      */
     public static function create(ClientInterface $client, CacheInterface $cache = null)
     {
