+++ tests/Service/Channel/ChannelServiceOpeningHoursMonthTest.php
@@ -1,65 +1,66 @@
 <?php
 
-namespace StadGent\Services\Test\OpeningHours;
+namespace StadGent\Services\Test\OpeningHours\Service\Channel;
 
-use StadGent\Services\OpeningHours\ChannelOpeningHoursService;
-use StadGent\Services\OpeningHours\Request\Channel\OpeningHoursPeriodRequest;
+use StadGent\Services\OpeningHours\Service\Channel\ChannelOpeningHoursService;
+use StadGent\Services\OpeningHours\Request\Channel\OpeningHoursMonthRequest;
 use StadGent\Services\OpeningHours\Response\OpeningHoursResponse;
 use StadGent\Services\OpeningHours\Value\OpeningHours;
+use StadGent\Services\Test\OpeningHours\Service\ServiceTestBase;
 
 /**
- * Tests for ChannelService::openingHoursPeriodHtml Method.
+ * Tests for ChannelService::openingHoursMonthHtml Method.
  *
  * @package StadGent\Services\Test\OpeningHours
  */
-class ChannelServiceOpeningHoursPeriodTest extends ServiceTestBase
+class ChannelServiceOpeningHoursMonthTest extends ServiceTestBase
 {
     /**
      * Test the return object.
      */
-    public function testOpeningHoursPeriod()
+    public function testOpeningHoursMonth()
     {
         $openingHours = $this->createOpeningHours();
         $client = $this->createClientForOpeningHours($openingHours);
 
         $channelService = new ChannelOpeningHoursService($client);
-        $responseOpeningsHours = $channelService->getPeriod(10, 20, '2020-01-02', '2020-02-02');
+        $responseOpeningsHours = $channelService->getMonth(10, 20, '2020-01-02');
         $this->assertSame($openingHours, $responseOpeningsHours);
     }
 
     /**
      * Test the return object from cache.
      */
-    public function testOpeningHoursPeriodFromCache()
+    public function testOpeningHoursMonthFromCache()
     {
         $openingHours = $this->createOpeningHours();
         $client = $this->createClientForOpeningHours($openingHours);
         $cache = $this->getFromCacheMock(
-            'OpeningHours:ChannelOpeningHoursService:period:10:20:2020-01-02:2020-02-02',
+            'OpeningHours:ChannelOpeningHoursService:month:10:20:2020-01-02',
             $openingHours
         );
 
         $channelService = new ChannelOpeningHoursService($client);
         $channelService->setCacheService($cache);
-        $responseOpeningHours = $channelService->getPeriod(10, 20, '2020-01-02', '2020-02-02');
+        $responseOpeningHours = $channelService->getMonth(10, 20, '2020-01-02');
         $this->assertSame($openingHours, $responseOpeningHours);
     }
 
     /**
      * Test the setCache when not yet cached.
      */
-    public function testOpeningHoursPeriodSetCache()
+    public function testOpeningHoursMonthSetCache()
     {
         $openingHours = $this->createOpeningHours();
         $client = $this->createClientForOpeningHours($openingHours);
         $cache = $this->getSetCacheMock(
-            'OpeningHours:ChannelOpeningHoursService:period:12:34:2020-01-02:2020-02-02',
+            'OpeningHours:ChannelOpeningHoursService:month:12:34:2020-01-02',
             $openingHours
         );
 
         $channelService = new ChannelOpeningHoursService($client);
         $channelService->setCacheService($cache);
-        $channelService->getPeriod(12, 34, '2020-01-02', '2020-02-02');
+        $channelService->getMonth(12, 34, '2020-01-02');
     }
 
     /**
@@ -129,7 +130,7 @@ class ChannelServiceOpeningHoursPeriodTest extends ServiceTestBase
     protected function createClientForOpeningHours(OpeningHours $openingHours)
     {
         $response = new OpeningHoursResponse($openingHours);
-        $expectedRequest = OpeningHoursPeriodRequest::class;
+        $expectedRequest = OpeningHoursMonthRequest::class;
         return $this->getClientMock($response, $expectedRequest);
     }
 }
